Fix a bunch of stuff

+++ commands/Utility/quote.js
@@ -1,8 +1,5 @@
 const { errorColor, successColor, neutralColor } = require("../../config.json")
 const Discord = require("discord.js")
-const { GoogleSpreadsheet } = require("google-spreadsheet")
-const creds = { "type": process.env.type, "project_id": process.env.project_id, "private_key_id": process.env.private_key_id, "private_key": process.env.private_key.replace(/\\n/gm, "\n"), "client_email": process.env.client_email, "client_id": process.env.client_id, "auth_uri": process.env.auth_uri, "token_uri": process.env.token_uri, "auth_provider_x509_cert_url": process.env.auth_provider_x509_cert_url, "client_x509_cert_url": process.env.client_x509_cert_url }
-const quotesSheet = process.env.quotes
 const { getDb } = require("../../lib/mongodb")
 
 module.exports = {
@@ -48,20 +45,20 @@ module.exports = {
                     .setFooter(executedBy, message.author.displayAvatarURL())
                 message.channel.stopTyping()
                 message.channel.send(embed)
-            } else addQuote(executedBy, message, strings, quote, author, collection)
-        } else if (args[0] === "edit") editQuote(executedBy, message, strings, args, collection)
-        else if (args[0] === "delete") deleteQuote(executedBy, message, strings, args, collection)
+            } else addQuote(executedBy, message, quote, author, collection)
+        } else if (args[0] === "edit") editQuote(executedBy, message, args, collection)
+        else if (args[0] === "delete") deleteQuote(executedBy, message, args, collection)
         else findQuote(executedBy, message, strings, args, collection)
     }
 }
 
 async function findQuote(executedBy, message, strings, args, collection) {
 
-    const all = collection.find({}).toArray()
+    const all = await collection.find({}).toArray()
 
     let quoteId
-    if (!args[0]) quoteId = Math.floor(Math.random() * Math.floor(all.length)) //generate random 0-base index number if no arg is given
-    else quoteId = args[0] //subtract 1 from argument in order to create 0-base index number
+    if (!args[0]) quoteId = Math.ceil(Math.random() * Math.floor(all.length)) //generate random id if no arg is given
+    else quoteId = Number(args[0])
 
     const quote = await collection.findOne({ id: quoteId })
     if (!quote) {
@@ -78,13 +75,13 @@ async function findQuote(executedBy, message, strings, args, collection) {
         .setColor(successColor)
         .setAuthor(strings.moduleName)
         .setTitle(quote.quote)
-        .setDescription(`      - <@!${quote.author}>`)
+        .setDescription(`      - ${quote.author}`)
         .setFooter(executedBy, message.author.displayAvatarURL())
     message.channel.stopTyping()
     return message.channel.send(embed)
 }
 
-async function addQuote(executedBy, message, strings, quote, author, collection) {
+async function addQuote(executedBy, message, quote, author, collection) {
 
     const all = collection.find({}).toArray()
     const quoteId = all.length + 1
@@ -92,7 +89,7 @@ async function addQuote(executedBy, message, strings, quote, author, collection)
     collection.insertOne({ id: quoteId, quote: quote, author: author }).then(result => {
         const embed = new Discord.MessageEmbed()
             .setColor(successColor)
-            .setAuthor(strings.moduleName)
+            .setAuthor("Quote")
             .setTitle("Success! The following quote has been added:")
             .setDescription(result.quote)
             .addFields(
@@ -105,16 +102,17 @@ async function addQuote(executedBy, message, strings, quote, author, collection)
     })
 }
 
-async function editQuote(executedBy, message, strings, args, collection) {
+async function editQuote(executedBy, message, args, collection) {
 
     const quoteId = args[1]
     args.splice(0, 2)
     const newQuote = args.join(" ")
+    if (!quoteId || !newQuote) throw "noQuote"
     const oldQuote = collection.findOne({ id: quoteId })
     collection.updateOne({ id: quoteId }, { $set: { quote: newQuote } })
     const embed = new Discord.MessageEmbed()
         .setColor(successColor)
-        .setAuthor(strings.moduleName)
+        .setAuthor("Quote")
         .setTitle(`Successfully edited quote #${quoteId}`)
         .addFields(
             { name: "Old quote", value: oldQuote.quote },
@@ -125,14 +123,15 @@ async function editQuote(executedBy, message, strings, args, collection) {
     message.channel.send(embed)
 }
 
-async function deleteQuote(executedBy, message, strings, args, collection) {
+async function deleteQuote(executedBy, message, args, collection) {
 
     const quoteId = args[1]
+    if (!quoteId) throw "noQuote"
     const oldQuote = collection.findOne({ id: quoteId })
     collection.deleteOne({ id: quoteId }, { quote: newQuote })
     const embed = new Discord.MessageEmbed()
         .setColor(successColor)
-        .setAuthor(strings.moduleName)
+        .setAuthor("Quote")
         .setTitle(`Successfully deleted quote #${quoteId}`)
         .addFields(
             { name: "User", value: oldQuote.author },
